[xabt] Combine post-trimming steps into single PostTrimmingPipeline task
Replace standalone AddKeepAlives and StripEmbeddedLibraries MSBuild tasks
with a single PostTrimmingPipeline task that opens assemblies once (via
DirectoryAssemblyResolver with ReadWrite) and runs both modifications in
a single pass. Extract StripEmbeddedLibrariesStep as an
IAssemblyModifierPipelineStep for reuse.
